Add files via upload
HTML file that can be used to monitor DUCO value in digital wallet, real-time miner monitoring, 3D animation. Modern and intuitive appearance. During operation, the file consumes moderate resources because the 3D animation runs continuously. Not recommended for mobile phones.
Change the end of the link in the HTML code that contains "users/******" with your username to retrieve the information from the server about the miner and the value of the wallet. Do not enter passwords or phrases from the wallet if you have them!!!